feat: Update Razorpay integration in Buy.tsx and BuyProduct.tsx
